+++ sphinx-docs/source/conf.py
@@ -203,8 +203,8 @@ try:
 except ImportError:
     pass  # dotenv not required for documentation build
 
-release = '0.9.1'
-version = '0.9.1'
+release = '0.9.2'
+version = '0.9.2'
 
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration